Move the shell expression to an appendix
The derivation is not directly relevant to the error analysis because we
don't mention the expressions in the rest of the text. I moved this to
the appendix so it won't interrupt the flow of the text. Being in the
appendix doesn't make it less important or new. It just serves to make
the text for fluid.