Edit the D-delta search section
Removed a lot of text from this section. As I was reading through it, I
didn't find the explanation about the different sections of the grid
necessary. In fact, it made the text more complicated than it needed to
be. The justification that we need was only:

1. We want the lowest D and highest delta
2. Anything out of the gray line is ruled out
3. D_linear is the best D for the exponential case so we can fix it to
   that.

This is pretty much what that whole explanation was about. Since we
don't need to explain the regions anymore, we don't really need the
extra figure, which makes the paper shorter and easier to follow.

Sorry, I think I asked Santiago to write it only to cut it out later.