Move the implementation section to Methodology
The section is small and describes the basis for the method developed,
so it fits in with the methodology. Included a link to the Github repo
(Zenodo DOI will come later).
Removed the details about the Fatiando code using numba. That would be
good on a software paper but not here.