Finished editing the conclusion
There was too much recap in the conclusion so I removed and summarized
some of it. I took out much of the algorithm and error analysis
description because it was repeated elsewhere. The conclusion should
give only a brief recap of the main results (in our case, the error
analysis). That's why I finished the conclusion with the results from
error analysis and a couple of sentences about why we can generalize
this to other density functions.