Edits to the introduction
I re-worded some sections to make things clearer and included more
citations to Lin and Denker. I re-wrote the part justifying why we need
a new method. The accuracy of the integration is probably not an issue
if you did a piecewise linear function (as our own results suggest).
Even using a high-degree polynomial wouldn't be too bad if the fit to
the density data is good. The main problem is that piecewise linear
raises the question of "what is the discretization interval" and "how
can we do this automatically"? We answer these with our algorithm. The
high-degree polynomial case would be very unstable in extrapolating the
density beyond the density profile data so it could cause problems for
gravity inversions.
Our method is automatic, which is the main selling point.