Move the description of grids and shells to the top
The description of each grid and shell model was in the linear density
section but we them for the other functions as well. I moved them up to
the beginning of the section and added some more justification about
why we chose these values (as requested by the reviewers).
I also simplified the formatting of the shell models table. I think this
is simpler and the caption can be a bit shorter because of it.