Edit the discussion
Rename to just Discussion. Remove some of the things repeated in the
intro about Lin2018 and Fukushima2018. I kept the part that mentions
that we could use the density discretization combined with piecewise
linear or polynomial functions (citing both papers).
Added some parts about why our method is independent of the density
function. Will reiterate this in the conclusions as well.
Argue why we can probably extrapolate the results here to any realistic
density function.