Edit the conclusions
Emphasize that the method is general purpose. We sacrifice speed for
usability and accuracy.
Added some more speculation about using the density-based discretization
with other methods.
Talk a bit about the Neuquen application and how the errors we see there
would translate into bad gravity inversions.

I removed the Gzz part because it felt very out of place. We don't talk
about it anywhere and have nothing to show. The only people who would
ask are the reviewers so we should include that in our response instead.