Replace "linear density" for "straight line"
On the density-based discretization algorithm we use a "straight line"
for computing the new discretization depth (if it's needed).
Naming this "straight line" as "linear density" could cause confusions
with the "linear density" used in the ratii determinations and may
induce readers to think that we are applying a piecewise linear
approximation of the density function.